sync: replace workspace with upstream MuseScore 4.6.5

+++ src/engraving/compat/dummyelement.h
@@ -25,7 +25,7 @@
 #include "../dom/engravingitem.h"
 
 namespace mu::engraving {
-enum class Pid : short;
+enum class Pid : int;
 
 class RootItem;
 }
